pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finreview"
version = "1.0.0"
description = "Automated financial analytical review pipeline for P&L, AR, and AP"
authors = [{name = "Financial Analytics Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",
    "python-pptx>=0.6.21",
    "matplotlib>=3.7.0",
    "plotly>=5.14.0",
    "streamlit>=1.28.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "structlog>=23.1.0",
    "pydantic>=2.0.0",
    "pyarrow>=12.0.0",
]

[project.optional-dependencies]
forecasting = [
    "prophet>=1.1.4",
    "pmdarima>=2.0.3",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "mypy>=1.4.0",
    "ruff>=0.0.282",
]

[project.scripts]
fin-review = "fin_review.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fin_review*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=fin_review --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

